create parse input file
